+++ app/controllers/admin/conversations/blocks_controller.rb
@@ -1,0 +1,9 @@
+module Admin
+  module Conversations
+    class BlocksController < ApplicationController
+      def index
+        @conversations = Conversation.blocked
+      end
+    end
+  end
+end