+++ test/components/expired_subscription_component_test.rb
@@ -2,7 +2,7 @@ require "test_helper"
 
 class ExpiredSubscriptionComponentTest < ViewComponent::TestCase
   test "renders the content if the business has an active subscription" do
-    user = users(:with_business_conversation)
+    user = users(:subscribed_business)
     render_inline ExpiredSubscriptionComponent.new(user, business: user.business) do
       "Content"
     end
@@ -10,8 +10,8 @@ class ExpiredSubscriptionComponentTest < ViewComponent::TestCase
   end
 
   test "renders the notice if the business does not have an active subscription" do
-    user = users(:with_business_conversation)
-    pay_subscriptions(:two).update!(ends_at: Date.yesterday)
+    user = users(:subscribed_business)
+    pay_subscriptions(:full_time).update!(ends_at: Date.yesterday)
 
     render_inline ExpiredSubscriptionComponent.new(user, business: user.business) do
       "Content"
@@ -21,8 +21,8 @@ class ExpiredSubscriptionComponentTest < ViewComponent::TestCase
   end
 
   test "renders the content if the user isn't associated with the business" do
-    user = users(:with_developer_conversation)
-    business = businesses(:with_conversation)
+    user = users(:prospect_developer)
+    business = businesses(:subscriber)
 
     render_inline ExpiredSubscriptionComponent.new(user, business:) do
       "Content"