+++ test/models/email_digests/new_developers_test.rb
@@ -3,10 +3,15 @@ require "test_helper"
 class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
   include DevelopersHelper
 
+  setup do
+    @business = businesses(:subscriber)
+  end
+
   test "daily emails for developers who signed up yesterday" do
     create_developers
+    @business.update!(developer_notifications: :daily)
 
-    args = {business: businesses(:two), developers: [@yesterday]}
+    args = {business: @business, developers: [@yesterday]}
     assert_enqueued_email_with BusinessMailer, :developer_profiles, args: do
       EmailDigests::NewDevelopers.new.send_daily_digest
     end
@@ -14,6 +19,7 @@ class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
 
   test "do not send daily emails if no new developers signed up yesterday" do
     create_developer(2.days.ago)
+    @business.update!(developer_notifications: :daily)
 
     assert_no_emails do
       EmailDigests::NewDevelopers.new.send_daily_digest
@@ -23,9 +29,10 @@ class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
   test "weekly emails for developers who signed up last week" do
     travel_to monday do
       create_developers
+      @business.update!(developer_notifications: :weekly)
 
       developers = [@yesterday, @two_days_ago, @seven_days_ago]
-      args = {business: businesses(:with_conversation), developers:}
+      args = {business: @business, developers:}
       assert_enqueued_email_with BusinessMailer, :developer_profiles, args: do
         EmailDigests::NewDevelopers.new.send_weekly_digest
       end
@@ -35,6 +42,7 @@ class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
   test "weekly emails are only sent on Mondays" do
     travel_to monday + 1.day do
       create_developers
+      @business.update!(developer_notifications: :weekly)
 
       assert_no_emails do
         EmailDigests::NewDevelopers.new.send_weekly_digest
@@ -45,6 +53,7 @@ class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
   test "do not send weekly emails if no new developers signed up this week" do
     travel_to monday do
       create_developer(8.days.ago)
+      @business.update!(developer_notifications: :weekly)
 
       assert_no_emails do
         EmailDigests::NewDevelopers.new.send_weekly_digest
@@ -54,7 +63,7 @@ class EmailDigests::NewDevelopersTest < ActionMailer::TestCase
 
   test "businesses without an active subscription don't receive emails, even if subscribed" do
     create_developers
-    businesses(:two).user.subscriptions.update_all(ends_at: Date.yesterday)
+    businesses(:subscriber).user.subscriptions.update_all(ends_at: Date.yesterday)
 
     assert_no_emails do
       EmailDigests::NewDevelopers.new.send_daily_digest