+++ test/system/message_submit_test.rb
@@ -3,15 +3,13 @@ require "application_system_test_case"
 class MessageSubmitTest < ApplicationSystemTestCase
   include Devise::Test::IntegrationHelpers
 
-  setup do
-    @user = users(:with_developer_conversation)
-    @conversation = @user.conversations.first
-  end
-
   test "submitting a conversation message using cmd+Enter" do
-    sign_in(@user)
+    user = users(:prospect_developer)
+    conversation = conversations(:one)
 
-    visit conversation_path(id: @conversation.id)
+    sign_in(user)
+
+    visit conversation_path(id: conversation.id)
     fill_in "message_body", with: "This is a cmd+Enter submission"
     find("#message_body").send_keys([:command, :enter])