+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### April
 
-* April 6 - Migrate email sending from Mailgun to Postmark - #365
+* April 7 - Migrate email sending from Mailgun to Postmark - #365
+* April 7 - Added `/admin/conversations` - #368
 
 ### March